backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.12.0
gunicorn==23.0.0
openai
